Make new stubs actual simple stubs
